third_party/blink/web_tests/platform/mac-mac11/external/wpt/performance-timeline/tentative/include-frames-originA-AB-expected.txt

This is a testharness.js-based test.
FAIL GetEntries of a document of origin A and its two child frames of origin A and B respectively. assert_equals: Total entries should 6. expected 6 but got 8
Harness: the test ran to completion.
